+++ src/world/walk.rs
@@ -1,22 +1,22 @@
 pub struct WalkIntention {
-    x: i32,
-    y: i32,
+    x: f32,
+    y: f32,
 }
 
 impl WalkIntention {
-    pub fn new(x: i32, y: i32) -> Self {
+    pub fn new(x: f32, y: f32) -> Self {
         return Self { x, y };
     }
 
     pub fn is_neutral(&self) -> bool {
-        return self.x == 0 && self.y == 0;
+        return self.x == 0.0 && self.y == 0.0;
     }
 
-    pub fn x(&self) -> i32 {
+    pub fn x(&self) -> f32 {
         return self.x;
     }
 
-    pub fn y(&self) -> i32 {
+    pub fn y(&self) -> f32 {
         return self.y;
     }
 }
